fix: avoid circular import
